Enable -Wshorten-64-to-32 on Clang and GCC and fix warnings in our code
Also disable this warning for external libs like qhttp and HIDRemote
since we don't control them and don't want to maintain fixes just for
warnings.

+++ src/display/osx/DisplayManagerOSX.cpp
@@ -61,9 +61,9 @@ bool DisplayManagerOSX::initialize()
       CGDisplayModeRef displayMode =
       (CGDisplayModeRef)CFArrayGetValueAtIndex(m_osxDisplayModes[displayid], modeid);
 
-      mode->m_height = CGDisplayModeGetHeight(displayMode);
-      mode->m_width = CGDisplayModeGetWidth(displayMode);
-      mode->m_refreshRate = CGDisplayModeGetRefreshRate(displayMode);
+      mode->m_height = (int)CGDisplayModeGetHeight(displayMode);
+      mode->m_width = (int)CGDisplayModeGetWidth(displayMode);
+      mode->m_refreshRate = (float)CGDisplayModeGetRefreshRate(displayMode);
 
       CFStringRef pixEnc = CGDisplayModeCopyPixelEncoding(displayMode);
